% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIV.SHAMP-package.R
\docType{package}
\name{EpiModelHIV.SHAMP-package}
\alias{EpiModelHIV.SHAMP-package}
\alias{EpiModelHIV.SHAMP}
\title{HIV Transmission Dynamics among MSM and Heterosexuals}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIV.SHAMP\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2017-06-27\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModelHIV.SHAMP provides extensions to EpiModel HIV to allow for simulating HIV transmission
dynamics among heterosexual and MSM of up to 5 race catagories simultaniously along with immigration.
}
\keyword{het}
\keyword{msm}
\keyword{package}
