% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.shamp.R
\name{init_shamp}
\alias{init_shamp}
\title{Epidemic Model Initial Conditions}
\usage{
init_shamp(prev.B.f = 0.005, prev.BI.f = 0.005, prev.H.f = 0.005,
  prev.HI.f = 0.005, prev.W.f = 0.005, prev.B.msf = 0.005,
  prev.BI.msf = 0.005, prev.H.msf = 0.005, prev.HI.msf = 0.005,
  prev.W.msf = 0.5, prev.B.msm = 0.1, prev.BI.msm = 0.1,
  prev.H.msm = 0.1, prev.HI.msm = 0.1, prev.W.msm = 0.1,
  prev.B.msmf = 0.05, prev.BI.msmf = 0.05, prev.H.msmf = 0.05,
  prev.HI.msmf = 0.05, prev.W.msmf = 0.05, ...)
}
\arguments{
\item{...}{Additional arguments passed to function.}

\item{prev.B}{Initial disease prevalence among black MSM.}

\item{prev.W}{Initial disease prevalence among white MSM.}
}
\value{
A list object of class \code{init_shamp}, which can be passed to EpiModel
function \code{\link{netsim}}.
}
\description{
Sets the initial conditions for a stochastic epidemic models
             simulated with \code{\link{netsim}}.
}
\keyword{SHAMP}

