% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.shamp.R
\name{param_shamp}
\alias{param_shamp}
\title{Epidemic Model Parameters}
\usage{
param_shamp(race.method = 1, time.unit = 7, method = 1,
  last.neg.test.B.f.int = 301, last.neg.test.BI.f.int = 301,
  last.neg.test.H.f.int = 301, last.neg.test.HI.f.int = 301,
  last.neg.test.W.f.int = 30, last.neg.test.B.msf.int = 301,
  last.neg.test.BI.msf.int = 301, last.neg.test.H.msf.int = 301,
  last.neg.test.HI.msf.int = 301, last.neg.test.W.msf.int = 30,
  last.neg.test.B.msm.int = 301, last.neg.test.BI.msm.int = 301,
  last.neg.test.H.msm.int = 301, last.neg.test.HI.msm.int = 301,
  last.neg.test.W.msm.int = 301, last.neg.test.B.msmf.int = 301,
  last.neg.test.BI.msmf.int = 301, last.neg.test.H.msmf.int = 301,
  last.neg.test.HI.msmf.int = 301, last.neg.test.W.msmf.int = 301,
  mean.test.B.f.int = 301, mean.test.BI.f.int = 301,
  mean.test.H.f.int = 301, mean.test.HI.f.int = 301,
  mean.test.W.f.int = 30, mean.test.B.msf.int = 301,
  mean.test.BI.msf.int = 301, mean.test.H.msf.int = 301,
  mean.test.HI.msf.int = 301, mean.test.W.msf.int = 30,
  mean.test.B.msm.int = 301, mean.test.BI.msm.int = 301,
  mean.test.H.msm.int = 301, mean.test.HI.msm.int = 301,
  mean.test.W.msm.int = 301, mean.test.B.msmf.int = 301,
  mean.test.BI.msmf.int = 301, mean.test.H.msmf.int = 301,
  mean.test.HI.msmf.int = 301, mean.test.W.msmf.int = 301,
  testing.pattern = "memoryless", test.window.int = 21,
  tt.traj.B.f.prob = c(0.077, 0, 0.356, 0.567), tt.traj.BI.f.prob = c(0.077,
  0, 0.356, 0.567), tt.traj.H.f.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.HI.f.prob = c(0.077, 0, 0.356, 0.567), tt.traj.W.f.prob = c(0.077,
  0, 0.356, 0.567), tt.traj.B.msf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.BI.msf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.H.msf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.HI.msf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.W.msf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.B.msm.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.BI.msm.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.H.msm.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.HI.msm.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.W.msm.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.B.msmf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.BI.msmf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.H.msmf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.HI.msmf.prob = c(0.077, 0, 0.356, 0.567),
  tt.traj.W.msmf.prob = c(0.077, 0, 0.356, 0.567), tx.init.B.f.prob = 0.092,
  tx.init.BI.f.prob = 0.092, tx.init.H.f.prob = 0.092,
  tx.init.HI.f.prob = 0.092, tx.init.W.f.prob = 0.092,
  tx.init.B.msf.prob = 0.092, tx.init.BI.msf.prob = 0.092,
  tx.init.H.msf.prob = 0.092, tx.init.HI.msf.prob = 0.092,
  tx.init.W.msf.prob = 0.092, tx.init.B.msm.prob = 0.092,
  tx.init.BI.msm.prob = 0.092, tx.init.H.msm.prob = 0.092,
  tx.init.HI.msm.prob = 0.092, tx.init.W.msm.prob = 0.092,
  tx.init.B.msmf.prob = 0.092, tx.init.BI.msmf.prob = 0.092,
  tx.init.H.msmf.prob = 0.092, tx.init.HI.msmf.prob = 0.092,
  tx.init.W.msmf.prob = 0.092, tx.halt.B.f.prob = 0.0102,
  tx.halt.BI.f.prob = 0.0102, tx.halt.H.f.prob = 0.0102,
  tx.halt.HI.f.prob = 0.0102, tx.halt.W.f.prob = 0.0102,
  tx.halt.B.msf.prob = 0.0102, tx.halt.BI.msf.prob = 0.0102,
  tx.halt.H.msf.prob = 0.0102, tx.halt.HI.msf.prob = 0.0102,
  tx.halt.W.msf.prob = 0.0102, tx.halt.B.msm.prob = 0.0102,
  tx.halt.BI.msm.prob = 0.0102, tx.halt.H.msm.prob = 0.0102,
  tx.halt.HI.msm.prob = 0.0102, tx.halt.W.msm.prob = 0.0102,
  tx.halt.B.msmf.prob = 0.0102, tx.halt.BI.msmf.prob = 0.0102,
  tx.halt.H.msmf.prob = 0.0102, tx.halt.HI.msmf.prob = 0.0102,
  tx.halt.W.msmf.prob = 0.0102, tx.reinit.B.f.prob = 0.00066,
  tx.reinit.BI.f.prob = 0.00066, tx.reinit.H.f.prob = 0.00066,
  tx.reinit.HI.f.prob = 0.00066, tx.reinit.W.f.prob = 0.00066,
  tx.reinit.B.msf.prob = 0.00066, tx.reinit.BI.msf.prob = 0.00066,
  tx.reinit.H.msf.prob = 0.00066, tx.reinit.HI.msf.prob = 0.00066,
  tx.reinit.W.msf.prob = 0.00066, tx.reinit.B.msm.prob = 0.00066,
  tx.reinit.BI.msm.prob = 0.00066, tx.reinit.H.msm.prob = 0.00066,
  tx.reinit.HI.msm.prob = 0.00066, tx.reinit.W.msm.prob = 0.00066,
  tx.reinit.B.msmf.prob = 0.00066, tx.reinit.BI.msmf.prob = 0.00066,
  tx.reinit.H.msmf.prob = 0.00066, tx.reinit.HI.msmf.prob = 0.00066,
  tx.reinit.W.msmf.prob = 0.00066, max.time.off.tx.full.int = 520 * 7,
  max.time.on.tx.part.int = 52 * 15 * 7, max.time.off.tx.part.int = 520 * 7,
  vl.acute.rise.int = 45, vl.acute.peak = 6.886, vl.acute.fall.int = 45,
  vl.set.point = 4.5, vl.aids.onset.int = 520 * 7, vl.aids.int = 52 * 2 *
  7, vl.fatal = 7, vl.full.supp = 1.5, vl.part.supp = 3.5,
  full.supp.down.slope = 0.25, full.supp.up.slope = 0.25,
  part.supp.down.slope = 0.25, part.supp.up.slope = 0.25,
  b.B.f.rate = 0.001/7, b.BI.f.rate = 0.001/7, b.H.f.rate = 0.001/7,
  b.HI.f.rate = 0.001/7, b.W.f.rate = 0.001/7, b.B.m.rate = 0.001/7,
  b.BI.m.rate = 0.001/7, b.H.m.rate = 0.001/7, b.HI.m.rate = 0.001/7,
  b.W.m.rate = 0.001/7, birth.age = 18, b.method = "fixed",
  msm.frac = 0.03, msmf.frac = 0.1, URAI.prob = 0.0082 * 1.09,
  UIAI.prob = 0.0031 * 1.09, URVI.prob = 0.0082 * 1.09, UIVI.prob = 0.0031
  * 1.09, acute.rr = 6, circ.rr = 0.4, condom.rr = 0.295,
  disc.outset.main.B.f.prob = 0.685, disc.outset.main.BI.f.prob = 0.685,
  disc.outset.main.H.f.prob = 0.685, disc.outset.main.HI.f.prob = 0.685,
  disc.outset.main.W.f.prob = 0.685, disc.outset.main.B.msf.prob = 0.685,
  disc.outset.main.BI.msf.prob = 0.685, disc.outset.main.H.msf.prob = 0.685,
  disc.outset.main.HI.msf.prob = 0.685, disc.outset.main.W.msf.prob = 0.685,
  disc.outset.main.B.msm.prob = 0.685, disc.outset.main.BI.msm.prob = 0.685,
  disc.outset.main.H.msm.prob = 0.685, disc.outset.main.HI.msm.prob = 0.685,
  disc.outset.main.W.msm.prob = 0.685, disc.outset.main.B.msmf.prob = 0.685,
  disc.outset.main.BI.msmf.prob = 0.685,
  disc.outset.main.H.msmf.prob = 0.685,
  disc.outset.main.HI.msmf.prob = 0.685,
  disc.outset.main.W.msmf.prob = 0.685, disc.at.diag.main.B.f.prob = 1,
  disc.at.diag.main.BI.f.prob = 1, disc.at.diag.main.H.f.prob = 1,
  disc.at.diag.main.HI.f.prob = 1, disc.at.diag.main.W.f.prob = 1,
  disc.at.diag.main.B.msf.prob = 1, disc.at.diag.main.BI.msf.prob = 1,
  disc.at.diag.main.H.msf.prob = 1, disc.at.diag.main.HI.msf.prob = 1,
  disc.at.diag.main.W.msf.prob = 1, disc.at.diag.main.B.msm.prob = 1,
  disc.at.diag.main.BI.msm.prob = 1, disc.at.diag.main.H.msm.prob = 1,
  disc.at.diag.main.HI.msm.prob = 1, disc.at.diag.main.W.msm.prob = 1,
  disc.at.diag.main.B.msmf.prob = 1, disc.at.diag.main.BI.msmf.prob = 1,
  disc.at.diag.main.H.msmf.prob = 1, disc.at.diag.main.HI.msmf.prob = 1,
  disc.at.diag.main.W.msmf.prob = 1, disc.post.diag.main.B.f.prob = 0,
  disc.post.diag.main.BI.f.prob = 0, disc.post.diag.main.H.f.prob = 0,
  disc.post.diag.main.HI.f.prob = 0, disc.post.diag.main.W.f.prob = 0,
  disc.post.diag.main.B.msf.prob = 0, disc.post.diag.main.BI.msf.prob = 0,
  disc.post.diag.main.H.msf.prob = 0, disc.post.diag.main.HI.msf.prob = 0,
  disc.post.diag.main.W.msf.prob = 0, disc.post.diag.main.B.msm.prob = 0,
  disc.post.diag.main.BI.msm.prob = 0, disc.post.diag.main.H.msm.prob = 0,
  disc.post.diag.main.HI.msm.prob = 0, disc.post.diag.main.W.msm.prob = 0,
  disc.post.diag.main.B.msmf.prob = 0, disc.post.diag.main.BI.msmf.prob = 0,
  disc.post.diag.main.H.msmf.prob = 0, disc.post.diag.main.HI.msmf.prob = 0,
  disc.post.diag.main.W.msmf.prob = 0, disc.outset.pers.B.f.prob = 0.527,
  disc.outset.pers.BI.f.prob = 0.527, disc.outset.pers.H.f.prob = 0.527,
  disc.outset.pers.HI.f.prob = 0.527, disc.outset.pers.W.f.prob = 0.828,
  disc.outset.pers.B.msf.prob = 0.527, disc.outset.pers.BI.msf.prob = 0.527,
  disc.outset.pers.H.msf.prob = 0.527, disc.outset.pers.HI.msf.prob = 0.527,
  disc.outset.pers.W.msf.prob = 0.828, disc.outset.pers.B.msm.prob = 0.527,
  disc.outset.pers.BI.msm.prob = 0.527, disc.outset.pers.H.msm.prob = 0.527,
  disc.outset.pers.HI.msm.prob = 0.527, disc.outset.pers.W.msm.prob = 0.828,
  disc.outset.pers.B.msmf.prob = 0.527,
  disc.outset.pers.BI.msmf.prob = 0.527,
  disc.outset.pers.H.msmf.prob = 0.527,
  disc.outset.pers.HI.msmf.prob = 0.527,
  disc.outset.pers.W.msmf.prob = 0.828, disc.at.diag.pers.B.f.prob = 1,
  disc.at.diag.pers.BI.f.prob = 1, disc.at.diag.pers.H.f.prob = 1,
  disc.at.diag.pers.HI.f.prob = 1, disc.at.diag.pers.W.f.prob = 1,
  disc.at.diag.pers.B.msf.prob = 1, disc.at.diag.pers.BI.msf.prob = 1,
  disc.at.diag.pers.H.msf.prob = 1, disc.at.diag.pers.HI.msf.prob = 1,
  disc.at.diag.pers.W.msf.prob = 1, disc.at.diag.pers.B.msm.prob = 1,
  disc.at.diag.pers.BI.msm.prob = 1, disc.at.diag.pers.H.msm.prob = 1,
  disc.at.diag.pers.HI.msm.prob = 1, disc.at.diag.pers.W.msm.prob = 1,
  disc.at.diag.pers.B.msmf.prob = 1, disc.at.diag.pers.BI.msmf.prob = 1,
  disc.at.diag.pers.H.msmf.prob = 1, disc.at.diag.pers.HI.msmf.prob = 1,
  disc.at.diag.pers.W.msmf.prob = 1, disc.post.diag.pers.B.f.prob = 0,
  disc.post.diag.pers.BI.f.prob = 0, disc.post.diag.pers.H.f.prob = 0,
  disc.post.diag.pers.HI.f.prob = 0, disc.post.diag.pers.W.f.prob = 0,
  disc.post.diag.pers.B.msf.prob = 0, disc.post.diag.pers.BI.msf.prob = 0,
  disc.post.diag.pers.H.msf.prob = 0, disc.post.diag.pers.HI.msf.prob = 0,
  disc.post.diag.pers.W.msf.prob = 0, disc.post.diag.pers.B.msm.prob = 0,
  disc.post.diag.pers.BI.msm.prob = 0, disc.post.diag.pers.H.msm.prob = 0,
  disc.post.diag.pers.HI.msm.prob = 0, disc.post.diag.pers.W.msm.prob = 0,
  disc.post.diag.pers.B.msmf.prob = 0, disc.post.diag.pers.BI.msmf.prob = 0,
  disc.post.diag.pers.H.msmf.prob = 0, disc.post.diag.pers.HI.msmf.prob = 0,
  disc.post.diag.pers.W.msmf.prob = 0, disc.inst.B.f.prob = 0,
  disc.inst.BI.f.prob = 0, disc.inst.H.f.prob = 0,
  disc.inst.HI.f.prob = 0, disc.inst.W.f.prob = 0,
  disc.inst.B.msf.prob = 0, disc.inst.BI.msf.prob = 0,
  disc.inst.H.msf.prob = 0, disc.inst.HI.msf.prob = 0,
  disc.inst.W.msf.prob = 0, disc.inst.B.msm.prob = 0,
  disc.inst.BI.msm.prob = 0, disc.inst.H.msm.prob = 0,
  disc.inst.HI.msm.prob = 0, disc.inst.W.msm.prob = 0,
  disc.inst.B.msmf.prob = 0, disc.inst.BI.msmf.prob = 0,
  disc.inst.H.msmf.prob = 0, disc.inst.HI.msmf.prob = 0,
  disc.inst.W.msmf.prob = 0, circ.B.prob = 0.874, circ.BI.prob = 0.874,
  circ.H.prob = 0.918, circ.HI.prob = 0.918, circ.W.prob = 0.918,
  ccr5.B.f.prob = c(0, 0.034), ccr5.BI.f.prob = c(0, 0.034),
  ccr5.H.f.prob = c(0, 0.034), ccr5.HI.f.prob = c(0, 0.034),
  ccr5.W.f.prob = c(0.021, 0.176), ccr5.B.m.prob = c(0, 0.034),
  ccr5.BI.m.prob = c(0, 0.034), ccr5.H.m.prob = c(0, 0.034),
  ccr5.HI.m.prob = c(0, 0.034), ccr5.W.m.prob = c(0.021, 0.176),
  ccr5.heteroz.rr = 0.3, num.inst.ai.classes = 1,
  base.ai.main.B.rate = 0.17, base.ai.main.BI.rate = 0.26,
  base.ai.main.H.rate = 0.23, base.ai.main.HI.rate = 0.26,
  base.ai.main.W.rate = 0.23, base.ai.pers.B.rate = 0.11,
  base.ai.pers.BI.rate = 0.16, base.ai.pers.H.rate = 0.14,
  base.ai.pers.HI.rate = 0.14, base.ai.pers.W.rate = 0.14,
  base.vi.main.B.rate = 0.17, base.vi.main.BI.rate = 0.26,
  base.vi.main.H.rate = 0.23, base.vi.main.HI.rate = 0.26,
  base.vi.main.W.rate = 0.23, base.vi.pers.B.rate = 0.11,
  base.vi.pers.BI.rate = 0.16, base.vi.pers.H.rate = 0.14,
  base.vi.pers.HI.rate = 0.16, base.vi.pers.W.rate = 0.14, ai.scale = 1,
  vi.scale = 1, cond.main.B.prob.msm = 0.38, cond.main.BI.prob.msm = 0.1,
  cond.main.H.prob.msm = 0.15, cond.main.HI.prob.msm = 0.15,
  cond.main.W.prob.msm = 0.15, cond.pers.always.prob.msm = 0.216,
  cond.pers.B.prob.msm = 0.26, cond.pers.BI.prob.msm = 0.26,
  cond.pers.H.prob.msm = 0.26, cond.pers.HI.prob.msm = 0.26,
  cond.pers.W.prob.msm = 0.26, cond.inst.always.prob.msm = 0.326,
  cond.inst.B.prob.msm = 0.27, cond.inst.BI.prob.msm = 0.27,
  cond.inst.H.prob.msm = 0.27, cond.inst.HI.prob.msm = 0.27,
  cond.inst.W.prob.msm = 0.27, cond.always.prob.corr.msm = 0.5,
  cond.main.B.prob.het = 0.38, cond.main.BI.prob.het = 0.1,
  cond.main.H.prob.het = 0.15, cond.main.HI.prob.het = 0.15,
  cond.main.W.prob.het = 0.15, cond.pers.always.prob.het = 0.216,
  cond.pers.B.prob.het = 0.26, cond.pers.BI.prob.het = 0.26,
  cond.pers.H.prob.het = 0.26, cond.pers.HI.prob.het = 0.26,
  cond.pers.W.prob.het = 0.26, cond.inst.always.prob.het = 0.326,
  cond.inst.B.prob.het = 0.27, cond.inst.BI.prob.het = 0.27,
  cond.inst.H.prob.het = 0.27, cond.inst.HI.prob.het = 0.27,
  cond.inst.W.prob.het = 0.27, cond.always.prob.corr.het = 0.5,
  cond.rr.BB = 1, cond.rr.BW = 1, cond.rr.WW = 1,
  cond.diag.main.beta.msm = -0.67, cond.discl.main.beta.msm = -0.85,
  cond.diag.pers.beta.msm = -0.67, cond.discl.pers.beta.msm = -0.85,
  cond.diag.inst.beta.msm = -0.67, cond.discl.inst.beta.msm = -0.85,
  cond.diag.main.beta.het = -0.67, cond.discl.main.beta.het = -0.85,
  cond.diag.pers.beta.het = -0.67, cond.discl.pers.beta.het = -0.85,
  cond.diag.inst.beta.het = -0.67, cond.discl.inst.beta.het = -0.85,
  role.B.msm.prob = c(0.242, 0.321, 0.437), role.BI.msm.prob = c(0.242,
  0.321, 0.437), role.H.msm.prob = c(0.242, 0.321, 0.437),
  role.HI.msm.prob = c(0.242, 0.321, 0.437), role.W.msm.prob = c(0.242,
  0.321, 0.437), role.B.msmf.prob = c(0.242, 0.321, 0.437),
  role.BI.msmf.prob = c(0.242, 0.321, 0.437), role.H.msmf.prob = c(0.242,
  0.321, 0.437), role.HI.msmf.prob = c(0.242, 0.321, 0.437),
  role.W.msmf.prob = c(0.242, 0.321, 0.437), vv.iev.B.prob = 0.42,
  vv.iev.BI.prob = 0.56, vv.iev.H.prob = 0.56, vv.iev.HI.prob = 0.56,
  vv.iev.W.prob = 0.49, prep.start = Inf, prep.elig.model = "base",
  prep.class.prob = c(0.211, 0.07, 0.1, 0.619), prep.class.hr = c(1, 0.69,
  0.19, 0.05), prep.coverage = 0, prep.cov.method = "curr",
  prep.cov.rate = 1, prep.tst.int = 90, prep.risk.int = 182,
  prep.risk.reassess = TRUE, immig.depart.BI.f = 0, immig.depart.HI.f = 0,
  immig.depart.BI.m = 0, immig.depart.HI.m = 0, immig.return.BI.f = 0,
  immig.return.HI.f = 0, immig.return.BI.m = 0, immig.return.HI.m = 0,
  immig.aq.prob.BI.f = 0, immig.aq.prob.HI.f = 0, immig.aq.prob.BI.m = 0,
  immig.aq.prob.HI.m = 0, msm.aq.prob.B = 0, msm.aq.prob.BI = 0,
  msm.aq.prob.H = 0, msm.aq.prob.HI = 0, msm.aq.prob.W = 0, ...)
}
\arguments{
\item{testing.pattern}{Method for HIV testing, with options \code{"memoryless"}
for constant hazard without regard to time since previous test, or
\code{"interval"} deterministic fixed intervals.}

\item{test.window.int}{Length of the HIV test window period in days.}

\item{max.time.off.tx.full.int}{Number of days off treatment for a full
suppressor before onset of AIDS, including time before diagnosis.}

\item{max.time.on.tx.part.int}{Number of days on treatment for a
partial suppressor beofre onset of AIDS.}

\item{max.time.off.tx.part.int}{Nnumber of days off treatment for a
partial suppressor before onset of AIDS, including time before
diagnosis.}

\item{vl.acute.rise.int}{Number of days to peak viremia during acute
infection.}

\item{vl.acute.peak}{Peak viral load (in log10 units) at the height of acute
infection.}

\item{vl.acute.fall.int}{Number of days from peak viremia to set-point
viral load during the acute infection period.}

\item{vl.set.point}{Set point viral load (in log10 units).}

\item{vl.aids.onset.int}{Number of days to AIDS for a treatment-naive
patient.}

\item{vl.aids.int}{Duration of AIDS stage infection in days.}

\item{vl.fatal}{Viral load in AIDS at which death occurs.}

\item{vl.full.supp}{Log10 viral load at full suppression on ART.}

\item{vl.part.supp}{Log10 viral load at partial suppression on ART.}

\item{full.supp.down.slope}{For full suppressors, number of log10 units that
viral load falls per time step from treatment initiation or re-initiation
until the level in \code{vl.full.supp}.}

\item{full.supp.up.slope}{For full suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected
value.}

\item{part.supp.down.slope}{For partial suppressors, number of log10 units
that viral load falls per time step from treatment initiation or
re-initiation until the level in \code{vl.part.supp}.}

\item{part.supp.up.slope}{For partial suppressors, number of log10 units that
viral load rises per time step from treatment halting until expected value.}

\item{birth.age}{Age (in years) of new arrivals.}

\item{b.method}{Method for calculating the number of expected births at each
time step, with \code{"fixed"} based on the number of persons at the
initial time step and \code{"varying"} based on the current time step.}

\item{URAI.prob}{Probability of transmission for a man having unprotected
receptive anal intercourse with an infected man at set point viral
load.}

\item{UIAI.prob}{Probability of transmission for an uncircumcised man having
unprotected insertive anal intercourse with an infected man at set
point viral load.}

\item{acute.rr}{Relative risk of infection (compared to that predicted by
elevated viral load) when positive partner is in the acute stage.}

\item{circ.rr}{Relative risk of infection from insertive anal sex when the
negative insertive partner is circumcised.}

\item{condom.rr}{Relative risk of infection from anal sex when a condom is
used.}

\item{circ.B.prob}{Probablity that a black new arrival in the population
will be circumcised.}

\item{circ.W.prob}{Probablity that a white new arrival in the population
will be circumcised.}

\item{ccr5.heteroz.rr}{Relative risk of infection for men who are heterozygous
in the CCR5 mutation.}

\item{num.inst.ai.classes}{Number of quantiles into which men should be
divided in determining their levels of one-off anal intercourse.}

\item{ai.scale}{General relative scaler for all act rates for model
calibration.}

\item{cond.rr.BB}{Condom probability scaler for black-black partnerships for
model calibration purposes.}

\item{cond.rr.BW}{Condom probability scaler for black-white partnerships for
model calibration purposes.}

\item{cond.rr.WW}{Condom probability scaler for white-white partnerships for
model calibration purposes.}

\item{prep.start}{Time step at which the PrEP intervention should start.}

\item{prep.elig.model}{Modeling approach for determining who is eligible for
PrEP. Current options are limited to: \code{"all"} for all persons who
have never been on PrEP and are disease-susceptible.}

\item{prep.class.prob}{The probability of adherence class in non-adherent,
low adherence, medium adherence, or high adherence groups (from Liu).}

\item{prep.class.hr}{The hazard ratio for infection per act associated with each
level of adherence (from Grant).}

\item{prep.coverage}{The proportion of the eligible population who are start
PrEP once they become eligible.}

\item{prep.cov.method}{The method for calculating PrEP coverage, with options
of \code{"curr"} to base the numerator on the number of people currently
on PrEP and \code{"ever"} to base it on the number of people ever on
PrEP.}

\item{prep.cov.rate}{The rate at which persons initiate PrEP conditional on
their eligibility, with 1 equal to instant start.}

\item{prep.tst.int}{Testing interval for those who are actively on PrEP. This
overrides the mean testing interval parameters.}

\item{prep.risk.int}{Time window for assessment of risk eligibility for PrEP
in days.}

\item{prep.risk.reassess}{If \code{TRUE}, reassess eligibility for PrEP at
each testing visit.}

\item{...}{Additional arguments passed to the function.}

\item{last.neg.test.B.int}{Time range in days for last negative test for
black men.}

\item{mean.test.B.int}{Mean intertest interval in days for black MSM who test.}

\item{last.neg.test.W.int}{Time range in days for last negative test for
white men.}

\item{mean.test.W.int}{Mean intertest interval in days for white MSM who test.}

\item{tt.traj.B.prob}{Proportion of black MSM who enter one of four
testing/treatment trajectories: never test or treat, test and never
initiate treatment, test and treated with partial viral suppression,
and test and treated with full suppression.}

\item{tt.traj.W.prob}{Proportion of white MSM who enter into the four
testing/treatment trajectories, as defined above.}

\item{tx.init.B.prob}{Probability per time step that a black MSM who has
tested positive will initiate treatment.}

\item{tx.init.W.prob}{Probability per time step that a white MSM who has
tested positive will initiate treatment.}

\item{tx.halt.B.prob}{Probability per time step that a black MSM who is
currently on treatment will halt treatment.}

\item{tx.halt.W.prob}{Probability per time step that a white MSM who is
currently on treatment will halt treatment.}

\item{tx.reinit.B.prob}{Probability per time step that a black MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{tx.reinit.W.prob}{Probability per time step that a white MSM who is
not currently on treatment but who has been in the past will
re-initiate treatment.}

\item{b.B.rate}{Rate at which black MSM enter the population.}

\item{b.W.rate}{Rate at which white MSM enter the population.}

\item{disc.outset.main.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a main partnership.}

\item{disc.outset.main.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a main partnership.}

\item{disc.at.diag.main.B.prob}{Probability that a black MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.main.W.prob}{Probability that a white MSM already in a main
partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.main.B.prob}{Probability that an HIV-infected black MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.post.diag.main.W.prob}{Probability that an HIV-infected white MSM
in a main partnership will disclose his status, assuming he didn't
at the start of the partnership or at diagnosis.}

\item{disc.outset.pers.B.prob}{Probability that an HIV-infected black MSM will
disclose his status at the start of a casual partnership.}

\item{disc.outset.pers.W.prob}{Probability that an HIV-infected white MSM will
disclose his status at the start of a casual partnership.}

\item{disc.at.diag.pers.B.prob}{Probability that a black MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.at.diag.pers.W.prob}{Probability that a white MSM already in a
casual partnership will disclose at the time of diagnosis.}

\item{disc.post.diag.pers.B.prob}{Probability that an HIV-infected black MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.post.diag.pers.W.prob}{Probability that an HIV-infected white MSM
in a casual partnership will disclose his status, assuming he
didn't at the start of the partnership or at diagnosis.}

\item{disc.inst.B.prob}{Probability that an HIV-infected black MSM will
disclose his status to a one-off partner.}

\item{disc.inst.W.prob}{Probability that an HIV-infected white MSM will
disclose his status to a one-off partner.}

\item{ccr5.B.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among black MSM.}

\item{ccr5.W.prob}{Vector of length two of frequencies of the Delta 32
mutation (homozygous and heterozygous, respectively) in the CCR5 gene
among white MSM.}

\item{base.ai.main.BB.rate}{Expected coital frequency in black-black main
partnerships (acts per day).}

\item{base.ai.main.BW.rate}{Expected coital frequency in black-white main
partnerships (acts per day).}

\item{base.ai.main.WW.rate}{Expected coital frequency in white-white main
partnerships (acts per day).}

\item{base.ai.pers.BB.rate}{Expected coital frequency in black-black casual
partnerships (acts per day).}

\item{base.ai.pers.BW.rate}{Expected coital frequency in black-white casual
partnerships (acts per day).}

\item{base.ai.pers.WW.rate}{Expected coital frequency in white-white casual
partnerships (acts per day).}

\item{cond.main.BB.prob}{Probability of condom use in a black-black main
partnership.}

\item{cond.main.BW.prob}{Probability of condom use in a black-white main
partnership.}

\item{cond.main.WW.prob}{Probability of condom use in a white-white main
partnership.}

\item{cond.pers.always.prob}{Fraction of men in casual partnerships who always
use condoms in those partnerships.}

\item{cond.pers.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black casual partnerships.}

\item{cond.pers.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white casual partnerships.}

\item{cond.pers.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white casual partnerships.}

\item{cond.inst.always.prob}{Fraction of men in instant partnerships who always
use condoms in those partnerships.}

\item{cond.inst.BB.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-black one-off partnerships.}

\item{cond.inst.BW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a black-white one-off partnerships.}

\item{cond.inst.WW.prob}{Of men who are not consistent condom users, per-act
probability of condom use in a white-white one-off partnerships.}

\item{cond.always.prob.corr}{Correlation coefficient for probability of always
using condoms in both casual and one-off}

\item{cond.diag.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.main.beta}{Beta multiplier for the log odds of using a
condom in a main partnership if the HIV-infected man has disclosed.}

\item{cond.diag.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.pers.beta}{Beta multiplier for the log odds of using a
condom in a casual partnership if the HIV-infected man has disclosed
his status.}

\item{cond.diag.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has been
diagnosed.}

\item{cond.discl.inst.beta}{Beta multiplier for the log odds of using a
condom in a one-off partnership if the HIV-infected man has disclosed
his status.}

\item{vv.iev.BB.prob}{Probability that in a black-black partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.BW.prob}{Probability that in a black-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}

\item{vv.iev.WW.prob}{Probability that in a white-white partnership of
two versatile men, they will engage in intra-event versatility
("flipping") given that they're having AI.}
}
\value{
A list object of class \code{param_shamp}, which can be passed to
EpiModel function \code{netsim}.
}
\description{
Sets the epidemic parameters for stochastic network models
             simulated with \code{\link{netsim}} for EpiModelHIV
}
\keyword{shamp}

