% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.prep.shamp.R
\name{prep_shamp}
\alias{prep_shamp}
\title{PrEP Module for up to 5 race groups heterosexuals and MSM.}
\usage{
prep_shamp(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\description{
Module function for implementation and uptake of pre-exposure
             prophylaxis (PrEP) to prevent HIV infection.
}
\keyword{het}
\keyword{module}
\keyword{msm}

