% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.test.R
\name{size.test}
\alias{size.test}
\title{Size test for memory allocation.}
\usage{
size.test(dat, at)
}
\arguments{
\item{dat}{Master data list object of class \code{dat} containing networks,
individual-level attributes, and summary statistics.}

\item{at}{Current time step.}
}
\value{
This function returns the \code{dat} with no changes.
}
\description{
This module measures the size of the elements of dat for diagnostic purposes.
}
\details{
This only checks object size in dat and not specific memory allocation functions.
}
\keyword{diagnostics}
